Fixing eslint
