+++ frontend/src/components/tempLandingPage.tsx
@@ -1,23 +1,23 @@
-import axios from "axios"
-import {useNavigate} from 'react-router-dom';
-
+/* eslint-disable @typescript-eslint/no-misused-promises */
+import axios from 'axios'
+import { useNavigate } from 'react-router-dom'
 
 export const TempLandingPage = () => {
-    const navigate = useNavigate();
+    const navigate = useNavigate()
 
     const handleLogout = async () => {
         try {
-            await axios.post("http://localhost:8000/user/logout");
+            await axios.post('http://localhost:8000/user/logout')
         } catch (error) {
-            console.error(error);
+            console.error(error)
         }
-        navigate('/login');
-    };
+        navigate('/login')
+    }
 
     return (
         <div>
             <h1>WeWrap Home</h1>
             <button onClick={handleLogout}>Logout</button>
         </div>
-    );
-};
+    )
+}
